requirements.txt
Flask==2.3.2
requests==2.31.0
pycryptodomex==3.18.0
protobuf==4.23.4
urllib3==1.26.16
gunicorn==20.1.0